Store itf index as int32

+++ pkg/collector/corechecks/network-devices/cisco-sdwan/payload/cedge_interface_test.go
@@ -20,7 +20,7 @@ func TestCEdgeInterface(t *testing.T) {
 		namespace                string
 		itf                      client.CEdgeInterfaceState
 		expectedID               string
-		expectedIndex            int64
+		expectedIndex            int32
 		expectedIndexError       string
 		expectedSpeed            float64
 		expectedOperStatus       devicemetadata.IfOperStatus
